requirements.txt
pytelegrambotapi
pydolarvenezuela
python-dotenv
pytz